requirements.txt
absl-py==2.0.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
fonttools==4.46.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
hydra-core==1.3.2
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==4.1.0
protobuf==4.23.4
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tb-nightly==2.16.0a20231216
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tf-keras-nightly==2.16.0.dev2023121610
tifffile==2023.12.9
tomli==2.0.1
torch==1.9.1
torch-tb-profiler==0.4.3
torchaudio==0.9.1
torchvision==0.10.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.1
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0
